Isobands via matix and extent

This is purely a re-orientation of isoband it self, rather than having to construct a degenerate rectilinear data set, we just provide a matrix and optionally an extent. The extent scaling can be done independently if wanted.

xisobands(x, extent = NULL, lo = NULL, hi = NULL, nlevs = 12L)

Arguments

x

matrix of numeric

extent

xmin,xmax,ymin,ymax georeference of the x matrix

lo

low levels

hi

high levels

nlevs

number of levels in the case that neither hi nor lo provided

Value

and isobands, iso object

Details

Matrix orientation is 'raster', i.e. like rasterImage() and not like image() which is isoband is expecting.

Values for lor and/or hi may be provided or left out. If both are left out can use nlevs to get a quantile set of levels from the data.

Examples

xisobands(diag(10))
#> [1] 10 10
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`0:0`
#> $`0:0`$x
#> numeric(0)
#> 
#> $`0:0`$y
#> numeric(0)
#> 
#> $`0:0`$id
#> integer(0)
#> 
#> 
#> $`1:1`
#> $`1:1`$x
#>  [1] 0.5 0.5 0.5 0.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 5.5 5.5 5.5 5.5 4.5 4.5 4.5
#> [20] 4.5 5.5 5.5 5.5 5.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 8.5 8.5 8.5 8.5 2.5 2.5
#> [39] 2.5 2.5 8.5 8.5 8.5 8.5 3.5 3.5 3.5 3.5 2.5 2.5 2.5 2.5 7.5 7.5 7.5 7.5 7.5
#> [58] 7.5 7.5 7.5 3.5 3.5 3.5 3.5 9.5 9.5 9.5 9.5 4.5 4.5 4.5 4.5
#> 
#> $`1:1`$y
#>  [1] 9.5 9.5 9.5 9.5 8.5 8.5 8.5 8.5 8.5 8.5 8.5 8.5 4.5 4.5 4.5 4.5 5.5 5.5 5.5
#> [20] 5.5 4.5 4.5 4.5 4.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 1.5 1.5 1.5 1.5 7.5 7.5
#> [39] 7.5 7.5 1.5 1.5 1.5 1.5 6.5 6.5 6.5 6.5 7.5 7.5 7.5 7.5 2.5 2.5 2.5 2.5 2.5
#> [58] 2.5 2.5 2.5 6.5 6.5 6.5 6.5 0.5 0.5 0.5 0.5 5.5 5.5 5.5 5.5
#> 
#> $`1:1`$id
#>  [1]  1  1  1  1  2  2  2  2  3  3  3  3  4  4  4  4  5  5  5  5  6  6  6  6  7
#> [26]  7  7  7  8  8  8  8  9  9  9  9 10 10 10 10 11 11 11 11 12 12 12 12 13 13
#> [51] 13 13 14 14 14 14 15 15 15 15 16 16 16 16 17 17 17 17 18 18 18 18
#> 
#> 
#> $`1:1`
#> $`1:1`$x
#>  [1] 0.5 0.5 0.5 0.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 1.5 2.5 2.5 2.5 2.5 2.5 2.5 2.5
#> [20] 2.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 4.5 4.5 4.5 4.5 4.5 4.5 4.5 4.5 5.5 5.5
#> [39] 5.5 5.5 5.5 5.5 5.5 5.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 8.5 8.5 8.5 8.5 8.5
#> [58] 8.5 8.5 8.5 7.5 7.5 7.5 7.5 7.5 7.5 7.5 7.5 9.5 9.5 9.5 9.5
#> 
#> $`1:1`$y
#>  [1] 9.5 9.5 9.5 9.5 8.5 8.5 8.5 8.5 8.5 8.5 8.5 8.5 7.5 7.5 7.5 7.5 7.5 7.5 7.5
#> [20] 7.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 6.5 5.5 5.5 5.5 5.5 5.5 5.5 5.5 5.5 4.5 4.5
#> [39] 4.5 4.5 4.5 4.5 4.5 4.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 3.5 1.5 1.5 1.5 1.5 1.5
#> [58] 1.5 1.5 1.5 2.5 2.5 2.5 2.5 2.5 2.5 2.5 2.5 0.5 0.5 0.5 0.5
#> 
#> $`1:1`$id
#>  [1]  1  1  1  1  2  2  2  2  3  3  3  3  4  4  4  4  5  5  5  5  6  6  6  6  7
#> [26]  7  7  7  8  8  8  8  9  9  9  9 10 10 10 10 11 11 11 11 12 12 12 12 13 13
#> [51] 13 13 14 14 14 14 15 15 15 15 16 16 16 16 17 17 17 17 18 18 18 18
#> 
#> 
#> attr(,"class")
#> [1] "isobands" "iso"